app/requirements.txt
Django==4.1.2
pytz==2022.4
tzdata==2022.4
django-admin-interface==0.22.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.4
sentry-sdk==1.9.10
django-prometheus==2.2.0
django-ninja==0.22.2
router==0.1
requests==2.31.0